// typescript-projects/utility-backups/src20250506/interface/types.ts
export type TStatusDocsAccessData = {
  uid: string;
  /** hash of naira withdrawal tx, which also serves as tx doc ID */
  appTxHash: string;
  /** full path to the related naira_admin doc */
  nairaAdminDocPath: string;
};
// ...
export type TNairaAdminTxWriteData = {
  utilityStatus?: AppTxStatus;
};

export type TNairaMetadataUtilityWriteData = {
  utilityStatus?: AppTxStatus;
  /** ID of the utility payment tx we generate and supply to the provider */
  reference?: string;
  txTimestamp?: number;
  providerId?: ProviderId;
  txPath?: string;
};